+++ src/index.js
@@ -1,5 +1,4 @@
 // three.js
-export { DebugTilesRenderer } from './three/DebugTilesRenderer.js';
 export { TilesRenderer } from './three/TilesRenderer.js';
 export { B3DMLoader } from './three/loaders/B3DMLoader.js';
 export { PNTSLoader } from './three/loaders/PNTSLoader.js';
@@ -14,21 +13,11 @@ export * from './three/math/Ellipsoid.js';
 export * from './three/math/EllipsoidRegion.js';
 export * as GeoUtils from './three/math/GeoUtils.js';
 export * from './three/math/GeoConstants.js';
-export * from './three/renderers/GoogleTilesRenderer.js';
-export * from './three/renderers/CesiumIonTilesRenderer.js';
 
 // three.js controls
 export { GlobeControls } from './three/controls/GlobeControls.js';
 export { EnvironmentControls } from './three/controls/EnvironmentControls.js';
 export { CameraTransitionManager } from './three/controls/CameraTransitionManager.js';
-
-// three.js plugins
-export { CesiumIonAuthPlugin } from './three/plugins/CesiumIonAuthPlugin.js';
-export { GoogleCloudAuthPlugin } from './three/plugins/GoogleCloudAuthPlugin.js';
-export * from './three/plugins/DebugTilesPlugin.js';
-
-// common plugins
-export { ImplicitTilingPlugin } from './base/plugins/ImplicitTilingPlugin.js';
 
 // common
 export { TilesRendererBase } from './base/TilesRendererBase.js';